remove strict type, merge start/end span into 1 subscriber, close span on error and on message handled
