+++ src/DependencyInjection/Compiler/SetInstrumentationTypePass.php
@@ -28,13 +28,8 @@ class SetInstrumentationTypePass implements CompilerPassInterface
         if ($container->hasParameter('open_telemetry.instrumentation.messenger.type')) {
             $messengerInstrumentationType = $container->getParameter('open_telemetry.instrumentation.messenger.type');
 
-            if ($container->hasDefinition('open_telemetry.instrumentation.messenger.trace.start_span')) {
-                $container->getDefinition('open_telemetry.instrumentation.messenger.trace.start_span')
-                    ->addMethodCall('setInstrumentationType', [$messengerInstrumentationType]);
-            }
-
-            if ($container->hasDefinition('open_telemetry.instrumentation.messenger.trace.end_span')) {
-                $container->getDefinition('open_telemetry.instrumentation.messenger.trace.end_span')
+            if ($container->hasDefinition('open_telemetry.instrumentation.messenger')) {
+                $container->getDefinition('open_telemetry.instrumentation.messenger')
                     ->addMethodCall('setInstrumentationType', [$messengerInstrumentationType]);
             }
         }
