+++ src/DependencyInjection/OpenTelemetryExtension.php
@@ -283,8 +283,7 @@ final class OpenTelemetryExtension extends ConfigurableExtension
             $container->removeDefinition('open_telemetry.instrumentation.messenger.trace.transport');
             $container->removeDefinition('open_telemetry.instrumentation.messenger.trace.transport_factory');
             $container->removeDefinition('open_telemetry.instrumentation.messenger.trace.middleware');
-            $container->removeDefinition('open_telemetry.instrumentation.messenger.trace.start_span');
-            $container->removeDefinition('open_telemetry.instrumentation.messenger.trace.end_span');
+            $container->removeDefinition('open_telemetry.instrumentation.messenger');
         }
 
         $this->setTracingInstrumentationParams($container, 'messenger', $config, $isConfigEnabled);
