+++ src/Instrumentation/Symfony/Console/TraceableConsoleEventSubscriber.php
@@ -30,8 +30,11 @@ final class TraceableConsoleEventSubscriber implements EventSubscriberInterface,
      */
     private array $excludeCommands = [];
 
+    /**
+     * @var string[]
+     */
     private array $notSupportedCommands = [
-        'messenger:consume' // designed to run indefinitely
+        'messenger:consume', // designed to run indefinitely
     ];
 
     public function __construct(
@@ -60,6 +63,9 @@ final class TraceableConsoleEventSubscriber implements EventSubscriberInterface,
         ];
     }
 
+    /**
+     * @return class-string[]
+     */
     public static function getSubscribedServices(): array
     {
         return [TracerInterface::class];
@@ -182,7 +188,7 @@ final class TraceableConsoleEventSubscriber implements EventSubscriberInterface,
 
     private function isNotSupported(Command $command): bool
     {
-        return in_array($command->getName(), $this->notSupportedCommands);
+        return in_array($command->getName(), $this->notSupportedCommands, true);
     }
 
     private function isAttributeTraceable(Command $command): bool
