+++ src/OpenTelemetry/Context/Propagator/PropagatorFactory.php
@@ -1,26 +1,21 @@
 <?php
 
-declare(strict_types=1);
-
 namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\OpenTelemetry\Context\Propagator;
 
 use OpenTelemetry\API\Baggage\Propagation\BaggagePropagator;
 use OpenTelemetry\API\Trace\Propagation\TraceContextPropagator;
 use OpenTelemetry\Context\Propagation\MultiTextMapPropagator;
-use OpenTelemetry\Context\Propagation\NoopTextMapPropagator;
-use OpenTelemetry\Context\Propagation\TextMapPropagatorInterface;
-use OpenTelemetry\SDK\Registry;
 
 class PropagatorFactory
 {
     /**
-     * Default propagators from https://opentelemetry.io/docs/specs/otel/configuration/sdk-environment-variables/#general-sdk-configuration
+     * Default propagators from https://opentelemetry.io/docs/specs/otel/configuration/sdk-environment-variables/#general-sdk-configuration.
      */
     public static function createDefault(): MultiTextMapPropagator
     {
         return new MultiTextMapPropagator([
             BaggagePropagator::getInstance(),
-            TraceContextPropagator::getInstance()
+            TraceContextPropagator::getInstance(),
         ]);
     }
 }
